functions/requirements.txt
firebase-functions
firebase-admin
google-genai
google-auth
google-adk
httpx
pydantic
requests
python-dotenv
google-cloud-storage
google-cloud-firestore
Pillow>=10.0.0
fal-client
colormath
resend
openpyxl
